move styleComposition to be a core module, which it is
