+++ src/components/Card.tsx
@@ -1,5 +1,5 @@
+import { withStyleProps } from "../styleComposition"
 import { RosesSC } from "../types"
-import { withStyleProps } from "../util/styleComposition"
 
 export const Card: RosesSC = withStyleProps({
   name: "Card",
