+++ src/components/Layout.tsx
@@ -1,7 +1,6 @@
-import styled, { Interpolation, StyledComponent } from "@emotion/styled"
-
+import styled from "@emotion/styled"
+import { boxStyle as baseStyle, withStyleProps } from "../styleComposition"
 import { RosesSC } from "../types"
-import { boxStyle as baseStyle, withStyleProps } from "../util/styleComposition"
 
 export const Box: RosesSC = withStyleProps({
   name: "Box",
