+++ src/components/Typography.tsx
@@ -1,5 +1,5 @@
+import { withStyleProps } from "../styleComposition"
 import { RosesSC } from "../types"
-import { withStyleProps } from "../util/styleComposition"
 
 export const Text: RosesSC = withStyleProps("Text")
 export const Heading: RosesSC = withStyleProps("Heading")
